requirements.txt
colorama==0.4.5
discord.py==1.7.3
python-dotenv==0.20.0
python_telegram_bot==13.12
